+++ common/src/main/java/com/wynntils/functions/generic/StringFunctions.java
@@ -12,7 +12,6 @@ import com.wynntils.utils.MathUtils;
 import com.wynntils.utils.StringUtils;
 import com.wynntils.utils.type.CappedValue;
 import com.wynntils.utils.type.RangedValue;
-import java.text.SimpleDateFormat;
 import java.util.List;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
@@ -84,27 +83,6 @@ public class StringFunctions {
         public FunctionArguments.RequiredArgumentBuilder getRequiredArgumentsBuilder() {
             return new FunctionArguments.RequiredArgumentBuilder(
                     List.of(new Argument<>("timestamp", Number.class, null)));
-        }
-    }
-
-    public static class FormatDateAdvancedFunction extends GenericFunction<String> {
-        @Override
-        public String getValue(FunctionArguments arguments) {
-            long timestamp = arguments.getArgument("timestamp").getLongValue();
-            String format = arguments.getArgument("format").getStringValue();
-
-            try {
-                String date = new SimpleDateFormat(format).format(timestamp);
-                return date;
-            } catch (IllegalArgumentException e) {
-                return "Invalid Format";
-            }
-        }
-
-        @Override
-        public FunctionArguments.RequiredArgumentBuilder getRequiredArgumentsBuilder() {
-            return new FunctionArguments.RequiredArgumentBuilder(List.of(
-                    new Argument<>("timestamp", Number.class, null), new Argument<>("format", String.class, null)));
         }
     }
