+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -1916,16 +1916,16 @@
   "function.wynntils.generic.formatDate.argument.timestamp": "Timestamp to format",
   "function.wynntils.generic.formatDate.description": "Formats a timestamp to a string version.",
   "function.wynntils.generic.formatDate.name": "Format Date",
-  "function.wynntils.generic.formatDateAdvanced.argument.format": "Format pattern",
-  "function.wynntils.generic.formatDateAdvanced.argument.timestamp": "Timestamp to format",
-  "function.wynntils.generic.formatDateAdvanced.description": "Format a timestamp to a string using an DateTimeFormatter pattern",
-  "function.wynntils.generic.formatDateAdvanced.name": "Format Date Advanced",
   "function.wynntils.generic.formatDuration.argument.seconds": "Seconds to format",
   "function.wynntils.generic.formatDuration.description": "Formats seconds to a shorter version.",
   "function.wynntils.generic.formatDuration.name": "Format Duration",
   "function.wynntils.generic.formatRanged.argument.value": "Ranged value to format",
   "function.wynntils.generic.formatRanged.description": "Formats a ranged value to a shorter version.",
   "function.wynntils.generic.formatRanged.name": "Format Ranged Value",
+  "function.wynntils.generic.formatTimeAdvanced.argument.format": "Format pattern",
+  "function.wynntils.generic.formatTimeAdvanced.argument.timestamp": "Timestamp to format",
+  "function.wynntils.generic.formatTimeAdvanced.description": "Format a timestamp to a string using an DateTimeFormatter pattern",
+  "function.wynntils.generic.formatTimeAdvanced.name": "Format Date Advanced",
   "function.wynntils.generic.fromHex.argument.hex": "The hex code representing the desired color. The leading # can be omitted",
   "function.wynntils.generic.fromHex.description": "Returns a color value based on provided hex argument",
   "function.wynntils.generic.fromHex.name": "Color from hex string",
